refactor: handle target version when resolving snapshot (#4001)
# Description

This PR updates the `resove_snapshot` to take an optional target version
argument.

This function right now is only used to harmonise some APis and make it
less error prone when a table is not loaded with state. While the change
is benign in this PR, we do need it when creating the next table
provider.

Signed-off-by: Robert Pack <robstar.pack@gmail.com>