fix: decode path before lookup (#3976)
# Description
Paths in the json log are double encoded, when we go over
logical_file_view.path() the paths are decoded but when you access them
directly in the add_actions_table they aren't so we need to decode as
well here to be able to look up in the map properly.

Imho, the double encoding in the logs seems like a bug but seems to have
been prevalent for a long time and this is just a tape fix to get old
behavior back.

# Related Issue(s)
- closes https://github.com/delta-io/delta-rs/issues/3939

---------

Signed-off-by: Ion Koutsouris <15728914+ion-elgreco@users.noreply.github.com>
Co-authored-by: R. Tyler Croy <rtyler@brokenco.de>
Co-authored-by: Robert Pack <42610831+roeap@users.noreply.github.com>