refactor: towards lazier snapshots (#4010)
# Description

As we try to merge `Snapshot` and `EagerSnapshot` we need to normalize
on APIs that we can expose on both snapshots. The `LogDataView`
construct is something that assumes materialised log data, which is not
something we can expose on a lazy snapshot.

This PR extends `Snapshot` to expose `file_views` as the eager variant
and removes some call sites of `log_data`.

Signed-off-by: Robert Pack <robstar.pack@gmail.com>