chore: fix windows uri test (#4011)
# Description
So the test prior was wrong. Windows base URLs handle the trailing slash
on drive roots in an _interesting_ way. For example, `C:/` is the actual
root of the drive, where as `C:` is actually the current working
directory. So there should always be a trailing slash for base root
drives. The code already handles this correctly, I'm just updating the
test to reflect that.

# Documentation

https://stackoverflow.com/questions/23955968/windows-path-with-no-slash-after-drive-letter-and-colon-what-does-it-point-to

Signed-off-by: Stephen Carman <shcarman@gmail.com>