fix: null handling when extracting scalars (#4014)
# Description

Currently we do not validate values we access when extracting a value
from an array. This PR adds a call to `is_valid` before we create a new
scalar, creating a NULL scalar otherwise.

I also asked a friendly AI helper to write some testing code for the
various methods in the scalars module.

Signed-off-by: Robert Pack <robstar.pack@gmail.com>