feat: kernel based table scans
We introduce a new TableProvider implementation that leverages the kernel's
capabilities for supporting advanced table features. A key design goal is
to fully utilize kernels and datafusions capabilities and only implement
what is strictly necessary in our codebase.

Getting to a point where we can defer materializing the log to as late as
possible and only for the parts that are strictly necessary is a key goal for
performance and efficiency. As such, the table provider can work with both
Snapshot and EagerSnapshot implementations.

Signed-off-by: Robert Pack <robstar.pack@gmail.com>
Co-authored-by: Stephen Carman <hntd187@users.noreply.github.com>